Move MyApp to other file and add firebase_auth package
