+++ quiz_quest/pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
   freezed_annotation: ^2.4.1
   flutter_bloc: ^8.1.3
   circular_countdown_timer: ^0.2.3
+  firebase_auth: ^4.16.0
   
 dev_dependencies:
   flutter_test:
